machineLearning/api/flask/requirements.txt
Flask===0.12
redis==2.10.5
python-dotenv==0.9.1
